task#27 implement adding a new issue https://app.tmetric.com/#/tasks/118627/2918101 implemented dormancy controllers to update and remove problems Refactoring

+++ src/main/java/com/antont/petclinic/user/issues/IssuesRepository.java
@@ -9,8 +9,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface IssuesRepository extends JpaRepository<Issue, Long> {
+public interface IssuesRepository extends JpaRepository<Issue, Integer> {
 
     List<Issue> findByPetOwner(Optional<User> user, Pageable pageable);
+
     List<Issue> findByDoctor(Optional<User> user, Pageable pageable);
 }
